Move file existance check
Move checking for the existance of the xml model description file into the
SourceTree init method so that it will also be called if this is called from a
module instead of main. Use with-open-as convention for opening the xml file.
Change the name of the file handle to avoid shadowing the built-in keyword
file.