Reverse default checkout behavior to checkout all components
Reverse the default checkout behavior to be more user friendly for novice
users. The basic expectation for the tool is that running 'checkout_model' will
checkout all the code. It would be counter-intuitive to the user require an
extra argument to get that behavior.

Replace the 'all' command line option with 'required'. Default behavior is now
to checkout all components and sub-componenents specified in the model
description file. If an advanced user only wants the required subset of the
components, they should pass the 'required' command line option.