Get model updates working safely.
Add functionality to evalue the state of the model repositories and determine
if all managed repositories are safe to update. If any repo is not safe, then
we inform the user and exit gracefully.

Added new unit tests for evaluating safe update status.

Testing: all tests ok, manual testing of modified and clean repos.