Implement using '.' as a repository url.
Specifying '.' as the repository url in an external description indicates that
the user does not want the tool to interact with remotes. Setting the repo to
'.' and the branch to the currently checked out branch will preserve the repo
state while other repos are updated.

Stop using random test-id for system tests because they were too hard to debug.
Figured out how to use the test name.

Testing:
  python2/3 - make test - pass
  python2 - cesm/clm manual testing checkout, status - ok