Change checkout process from depth first to breadth first.
Change the order of the checkout process from depth first to breadth first.
This cleans up the output during the initial checkout of a container repository
by doing all checkouts at the current level before doing sub externals. The
disadvantage is that if a checkout dies, a component may be in an incomplete
state. For example, clm is checked out, but it's externals are not. Then
another component checkout fails. This should be easily fixed by a rerun of the
checkout.

Testing:
    make test - python2/3 - pass, 1 skip
    manual testing - cesm-demo, clm-demo, escomp/cesm - ok checkout, status