Ignore untracked files when running checkout
Untracked files in a git or svn external have the potential to abort a checkout
and leave the sandbox in a poorly defined state that the user needs to resolve.
The initial implementation refused to update if untracked files were present to
prevent this. This restriction has been deamed too restrictive. It is possible
to request svn to do a dry-run and determine if a checkout or switch will be
successful. Git does not have a mechanism to determine if a checkout will be
successful before running, and there are not any clear workarounds.

For now untracked files are being ignored. If problems arise, restrictions and
checks will be re-added to address specific problems.

Testing:
  new unit tests added for new git functionality
  existing tests were updated to test for new behavior.
  python2 - make test - all tests pass, one skip.