Abort if we find an svn external in dirty state following switch
Unlike git, we cannot rely on the return status here; instead, we need
to check 'svn status' following the switch.

This is to signal to the user if there was an problem in doing the
switch, such as a conflict in the final state. This can happen if you
had an untracked file in the working directory with the same name as a
tracked file in the new revision.