If we're already in sync, do not show output from the checkout command
However, specifying '--verbose --verbose' will still show output from
the checkout command even if we're already in sync.

This implementation feels like a bit of a hack. On the other hand, I can
convince myself that it makes sense as a general rule to decrease the
verbosity level of the checkout if we're in sync.