Update mixed-use test repo.
Add a sub-externals description file to the mixed-use container/external test
repo with a single sub-external that needs to be checked out. Make a branch in
the repo with a modified set of externals. Update system test using this repo
to not create a sub-externals file, and to only look for a single sub-external.

Testing:
  make test - python2 - pass, one skip.