Finish implementing system test for mixed-use externals
Finish implementing a system level test for mixed use externals as an external.
This removes the unittest.skip decorator from the test_container_full system
test.

Note: this test reveals a previously reported bug that when an external with
sub-externals gits updated, the sub-externals will not be updated because the
sub-external information contained in memory has gone stale.

Testing:
    make test - python2 - all tests pass, no skips.