Bugfix: stale sub-externals after checkout.
When an external has sub-externals, and the sub-externals description file is
modified by a checkout, the in-memory cached description of the sub-externals
is stale and potentially out date.

Change the logic when doing breadth-first search for sub-externals. Don't rely
on the in-memory sub-externals description, but delete it and re-read the
appropriate externals description file.

Update the test_container_full system test to check for the new behavior.

Testing:
    make test - python3 - all tests pass