Rewrite _current_ref to use plumbing rather than parsing porcelain
Previously, _current_ref attempted to do everything with the output from
git branch -vv. This required some complex regex matches that were
fragile and hard to maintain. In addition, differences between git v. 1
and git v. 2 meant that the result was incorrect in some cases -
particularly, in the case where we have "detached from foo" (which is
the text that always appeared for a detached head in v 1, but in v 2
means we are no longer at foo).

I have rewritten _current_ref to use various plumbing commands. This
also resulted in being able to simplify some of the unit tests.