Point to new refcases for fully-coupled compsets
The new refcases are the same as the old except for the CLM initial
condition file, which has been updated for compatibility with
clm5.0.dev006.

I have shown that these give bit-for-bit results with the old refcases
by comparing (1) B1850 and BW1850 runs done with clm5.0.dev005 with the
old refcases with (2) B1850 and BW1850 runs done with clm5.0.dev006 with
the new refcases.