Incorporate feedback from Gunter Leguy, and other minor edits
A significant change is clarifying the use of $VAR. I removed uses of
$VAR as things the user should fill in, because this seems confusing to
new users, who think this means that VAR should be a shell environment
variable. I have also repeated the description of the use of $VAR near
the top of the quickstart page, as Gunter suggested.

I have also added a bit of clarification about CASEROOT, as I did in the
cime documentation in https://github.com/ESMCI/cime/pull/2693.

I have also explicitly pointed out that most output won't appear under
the case directory, since this was confusing to Gunter.