Remove out-of-the-box use of refcases
It can be a pain for CLM to support refcases. Those present at today's
CSEG meeting don't see any reason why we need to support the cesm2.0
refcases in cesm2.2. So we'll remove these for now to decrease the cost
of making changes in CLM (and elsewhere), and will just bring back
refcases as we get closer to a cesm2.2 release.