Do NOT include all defaultio mods in the cmip6 testmods
Including all of the defaultio mods led to some of the cmip6 settings
being overridden by the defaultio settings for CLM and CICE (and this
would become true for CAM, too) - so that we were not actually testing
all of the cmip6 settings. We might have been able to handle this by
being careful about the order in which the cmip6 usermods vs. defaultio
were listed in the include_user_mods files. But CIME doesn't make
guarantees about which included user mods takes precedence (first or
last), so this would have been fragile. So we're going with the safer
approach of only including the bare minimum of defaultio mods (just for
POP, which doesn't pull in any mods via the cmip6 usermods directory).