Add B1850cmip6 compset, remove cmip6 usermods directories
The new plan is to use new cmip6-specific compsets that automatically
add the relevant components' usermods directories, rather than having
top-level usermods directories in the top-level cime_config.

Note that the new B1850cmip6 compset will only work with changes in CLM
and CICE that are currently on branches, in outstanding PRs for each of
these components.