Remove J1850G compset for now
This isn't working as written. Part of the problem is that the datm data
it points to aren't available. It could be updated to point to
CRUv7. However, that still leaves an issue of where to get the CO2
forcing data if we want this compset to replicate what's actually being
used for LIWG spinup work (which uses _BGC%BDRD). Rather than trying to
resolve this now, @katetc and I thought it best to just remove this
compset for the CESM2.1 release and then revisit this for a later
release.