Set CLM's use_init_interp=.false. for ERI f09_g17 B1850 tests
Use a new testmod for this. This testmod is designed for ERI tests of
compset-grid combinations that set CLM's use_init_interp = .true. by
default (which is done because this is needed for compatibility with the
hybrid refcase used for that compset-grid combination). The problem with
this setting is that, in the branch portion of the ERI test, CLM dies
with an error message if use_init_interp is true:

   ERROR: Can only set use_init_interp if finidat is set

(because finidat is NOT set in a branch run).

Currently, ERI tests do a startup run for the ref1 case, so we don't
actually need the use_init_interp setting that we would normally need
for this compset-grid combination. So it's safe to simply turn off
use_init_interp for all of the cases in this ERI test.

I and others involved in this discussion (Erik Kluzek and Chris Fischer)
don't love this solution of using a different testmod for these specific
ERI tests, but we weren't able to come up with any other solution that
we liked any better. Alternatives we considered and rejected were: (1)
Don't set init_interp in config_compsets, instead relying on a different
refcase (but this gets messy in terms of managing refcases, and it also
is less clear that two cases really use the same refcase but one uses
init_interp with that refcase); (2) Change the ERI test itself (or
introduce an extended ERI test) which sets use_init_interp = .false. in
user_nl_clm just for the branch case (but we don't like changing the
main ERI test, and allowing for easy extension would require refactoring
the main ERI test, which we don't want to do right now); (3) Removing or
changing the CLM error check, including possibly putting this
error-check at build-namelist time and turning it into a warning (but we
like having this be an error rather than just a warning).