Remove testing on mac os
This is failing. It's using python2, which is no longer supported
(though I don't know if that's why it's failing). I don't see the point
in trying to update it to use python3, since I feel we have sufficient
coverage with our linux testing, so for now I'm just removing it.