test_sys_checkout: use actual paths in on-the-fly configs rather than MANIC_TEST_BARE_REPO_ROOT env var. This will make it easier to test (in the near future) that checkout_externals actually checked out the desired repo dir.
The env-var path is still tested by two tests which rely on a checked-in externals.cfg that relies on an environment variable.

More clearly document that env vars are passed to checkout_externals via os.environ, and that the command line assembled in _execute_checkout_in_dir is just for manual reproducibility.