Change new tests to be at least a day
It didn't work to do 7-hour ERS tests: this led to failures in the
comparisons of some MOM output files. (I didn't look into this
carefully, but at a glance, it looks like it was comparing the file from
the initial time, I guess because that was the only one available.)

I want a debug ERS test of the standard configuration, so I'm keeping
that but lengthening it to 3 days. For the agrid and ogrid, it doesn't
seem worth the compute time to make these ERS tests, because I'm not
sure that there is any different restart-related code covered for these
compared with the standard configuration; so I'm changing those to 1-day
SMS tests. (It would probably work to make them less than a day, but
making them 1 day lets us output component history files and avoid
changing coupling frequencies for any components.)